feat: updated stories, cleared the code
