+++ src/shared/ui/selectBox/SelectBox.module.scss
@@ -1,26 +1,5 @@
-// .selectTrigger {
-// 	border: 1px solid var(--color-dark-100);
-// 	border-radius: 2px;
-// 	font-size: var(--font-size-m);
-// 	height: 36px;
-// 	background: none;
-// 	display: flex;
-// 	justify-content: space-between;
-// 	align-items: center;
-// 	padding: 6px 12px;
-// 	// width: 210px;
-// }
-
-
-// .selectContent {
-// 	border: 1px solid var(color--light-100);
-// 	// width: 210px;
-// }
-
-
 .trigger {
 	all: unset;
-
 	border: 1px solid var(--color-dark-100);
 	border-radius: 2px;
 	font-size: var(--font-size-m);
@@ -29,27 +8,32 @@
 	background: none;
 	display: flex;
 	justify-content: space-between;
-	gap: 10px;
+	gap: 15px;
 	align-items: center;
 	padding: 6px 12px;
 	cursor: pointer;
+
 	&:hover {
 		color: var(--color-light-900);
 	}
+
 	&:focus {
 		border-color: 2px solid var(--color-accent-500);
 		color: var(--color-light-900);
 	}
-	// &:active{
-	// 	background-color: var(--color-dark-500);
-	// 	border-color: var(--color-light-100);
-	// }
+
 	&[data-state='open'] {
 		background-color: var(--color-dark-500);
 		color: var(--color-light-100);
 		border-color: var(--color-light-100);
 		border-radius: 2px 2px 0 0;
+		.icon {
+				svg {
+				transform: rotate(-180deg);
+			}
+		}
 	}
+
 	&[data-disabled] {
 		border-color: var(--color-dark-100);
 		color: var(--color-dark-100);
@@ -60,10 +44,11 @@
 .icon {
 	height: 24px;
 	width: 24px;
+	
 	svg {
+		transition: all 0.3s ease 0s;
 		width: 100%;
 		height: 100%;
-		
 	}
 }
 
@@ -75,8 +60,6 @@
 	border-radius: 0 0 2px 2px;
 }
 
-
-
 .item {
 	font-size: var(--font-size-m);
 	color: var(--color-light-100);
@@ -87,36 +70,10 @@
 	width: 100%;
 	cursor: pointer;
 	user-select: none;
-	// position: relative;
-	// &[data-disabled] {
-		// pointer-events: none;
-	// 	color: var(--mauve-8);
-	// }
+
 	&[data-highlighted] {
 		outline: none; 
 		background-color: var(--color-dark-300);
 		color: var(--color-accent-500);
 	}
 }
-
-// .Label {
-// 	padding: 0 25px;
-// 	font-size: 12px;
-// 	line-height: 25px;
-// 	color: var(--mauve-11);
-// }
-
-// .Separator {
-// 	height: 1px;
-// 	background-color: var(--violet-6);
-// 	margin: 5px;
-// }
-
-// .ItemIndicator {
-// 	position: absolute;
-// 	left: 0;
-// 	width: 25px;
-// 	display: inline-flex;
-// 	align-items: center;
-// 	justify-content: center;
-// }
